Check vGPU/pGPU compatibility
This inserts a call to check vGPU/pGPU compatiblity based on metadata. The test passes when no metadata is available but carries out a real check when it is. The assumption is that this data is only available in a VM.resume context because vGPU metadata is only saved before VM.suspend.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>