CA-271874: Local-to-local storage migration using CLI, without specifying VDI-map, fails badly -rework++
The issue is regarding the way the default SR is chosen, when no VDI mapping is provided. The current behavior is the default SR of the destine pool will be chosen. However in case of none shared SR in pool and the SR is not on the destine host, there will be problem.

The fix will change the strategy to choose an SR, in that SRs will be chosing based on the PBD the destine host is attached, shared preferred.

Signed-off-by: Yarsin He <yarsin.he@citrix.com>