Test that Cluster_host.destroy is disallowed when SR is attached
If Xapi_cluster_host.destroy allows the operation, it will get past the
initial checks and will try to communicate with xapi-clusterd, which
will result in a different Failure from one we expect, and will cause
the unit test to fail. Conversely, if Xapi_cluster_host.destroy is
implemented correctly, it will raise the Failure we expect from the
initial checks, and will not proceed to communicate with xapi-clusterd,
and the unit test will pass. We will need to finalize the exceptions
and expect a more specific one to make this unit test more robust.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>