updates_test: do not rely on json as a string but compare the Rpc structure
This fixes a test that broke after the latest rpc update.
The new json serialization drops unneeded characters (e.g. pretty print newlines and spaces),
so tests that are based on specific output strings can break without any real breakage.

In this case the failure was due to:
```
got ->      {"updates":[{"id":1,"v":"[\"Foo\",\"a\"]"}],"barriers":[[1,2,[{"id":1,"v":"[\"Foo\",\"a\"]"}]]]}
expected -> {"updates": [{"id": 1, "v": "[\"Foo\", \"a\"]"}], "barriers": [[1, 2, [{"id": 1, "v": "[\"Foo\", \"a\"]"}]]]}
```

The two json representations are matching but we were relying on the
wrong invariant.
This change makes the test more robust with respect to this kind of
changes, although still not perfect (the values of v are still json
strings).

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>