Quicktest_cbt: don't fail in cleanup due to VDI refs becoming invalid
There is a race condition where we get the list of VDIs, but before we
process them some of them get removed from the DB, and the cleanup fails
with an invalid ref error.

This can happen for example when SM creates an internal VDI record in
the DB, and then later it removes it, in parallel with the cleanup at
the end of the CBT quicktest.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>