CA-275660 release vGPUs from VM on suspend (#3364)
When a VM is suspended, it must release all vGPUs it has allocated. So
far, this is only happens on shutdown but not suspend. This commit
rectifies this.

A vGPU stays attached to a pGPU when suspended. Hence, only reset
resident_on and scheduled_to_be_resident_on of the vGPU.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>