CA-272163: Redirect snapshot to host that has activated VDI (if any)
But only if the SM backends asks us to do it.
The SM backend redirects based on host UUID, lookup the UUID here and transform
it to an IP address, it should be a host from the same pool.

If this gets called during storage migration it should still work as long as
the VM was activate on a host in this same pool, but that is untested.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>