CP-24694: disable Cluster_host on shutdown and reenable on startup
XAPI replugs any unplugged PBDs on startup, so we don't need the 'soft' unplug.
Also if those plugs are to succeed then we need to reenable the Cluster_host for
the current host on startup before attempting to plug the PBD.

We also need to disable the Cluster_host in the DB on shutdown, otherwise
you cannot shutdown more than one host.

We can unconditionally call Cluster_host.disable, because it will now be able to
do the necessary prechecks on whether the PBD is plugged or not.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>