CA-276964 on import, set suspend_SR to null when unknown
On importing a VM, we must check that its suspend_SR is meaningful and
set it to Null if it isn't. The existing code has the test flipped.

Helpers.check_sr_exists ~__context ~self:vm_record.API.vM_suspend_SR

The above call returns `Some vM_suspend_SR` if it exists and `None`
otherwise. The test needs to look for None but previously did not.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>