Add new helper to get domain_type regardless of the power_state
Note that this raises an internal error if the found domain_type is
unspecified, so that we don't need to handle that case everywhere in the code.

For halted VMs, the VM.domain_type is relevant, because is what the domain will
be when the VM is started. This field should only transiently be unspecified,
while the VM is created, or imported/upgraded from an older version of xapi.

For other power states, we look at VM_metrics.current_domain_type, because the
VM field is R/W and changes will take effect only when the VM is restarted
(note: not rebooted ;)). The field is unspecified initially, but should never
be once the VM has completed its first start.

However, there appears to be a short window after the power_state is set to
running, while current_domain_type is still unspecified.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>