CA-279161 if unavailable, don't update pGPU compat data
On startup, xapi updates its GPU records. This includes pGPU
compatibility meta data. Obtaining this data might fail if the pGPU is
used in pass-trough mode at the time. In the past this lead to empty
meta data overwriting the existing meta data. Since a restart of xapi
can happen for many reasons, this led to losing pGPU meta data which was
needed later.

This commit prevents existing pGPU meta data being overwritten if no
meta data can be obtained. This could introduce stale data, so why is
this safe?

* Any xapi restart when pGPU are accessible will obtain the correct value.
  This provides a way to correct any potential problem.

* If the corresponding card is removed, a reboot is necessary and during
  xapi startup the pGPU object will be removed.

* If the corresponding card is replaced in the same PCI slot by a
  different card, a reboot is required during which pGPUs are
  accessible.  If the card does not provide pGPU metadata, we would not
  overwrite the existing data. However, provided the new card has a
  different PCI identifier, the old pGPU object will be removed entirely
  and a new object created.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>