Sync ISCSI iqn and multipathing config files at startup
We have to make sure these are handled correctly on startup, in case
there was a hard reboot and had to recover from a redo log, or when one
of these are in a nonpersistent file. We cannot know the state of these
config files before xapi started, so it's the safest to sync them at
startup.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>