CP-26298: Increase the max_stunnel in xapi
This commit is to increase the max_stunnel value when we support larger
number of hosts(64 etc) in a pool.

If max_stunnel is too small then sometimes a call from master to slave
will have higher latency than necessary, as the master will initiate
creation of a new TLS session instead of re-using an existing one.

Signed-off-by: Taoyong Ding <taoyong.ding@citrix.com>