test_cpuid_helpers: fix comparison tests
The unit tests that compare feature sets of different lengths were not
expecting the right outcome. The way `is_subset` and `is_strict_subset` are
meant to work is by zero-extending the shorter set before the comparison. The
implementation of these functions is not quite right in this regard, and
unfortunately, the unit tests matched the wrong behaviour.

Note that some of these tests will now fail. The following commit will fix the
`is_subset` and `is_strict_subset` functions.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>