CA-271867: Local-to-local storage migration fails with out-of-space error
It is found when the source SR is nearly exhausted and migration is invoked to some remote destine, the migration will fail but the message will be confusing as remote space is available and the actual problem is local space shortage to do some snapshot.
Besides, there might be case that the remote is left with some garbage vdi on the destine vdi.

This change  added a new alert specific for this kind of issue when local space is not available when doing snapshot at migration, and upon failure, remote the remote vdi for the snapshot mirror.

Signed-off-by: Yarsin He <yarsin.he@citrix.com>