VDI.snapshot, clone: pass down complete vdi_info
We use vdi_info_of_vdi_rec to ensure that the vdi_info we pass down to
SMAPIv2 contains the correct information about the VDI.
Previously when we were debugging an issue, we were looking at the
vdi_info printed out by SMAPIv2, and it was misleading as many of the
fields, such as is_a_snapshot and cbt_enabled, were incorrect because
they were just taken from the default_vdi_info.
I determine the snapshot time in SMAPIv2, to ensure that in the vdi_info
we only have to pass down the fields of the snapshotted VDI, not the new
fields of the snapshot. This means that SMAPIv3 plugins will also have
to set the snapshot time correctly, instead of relying on the old
behaviour.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>