CA-266936: Use a CArray instead of an ocaml Buffer
The original code was passing an ocaml managed Buffer as
a buffer for `pciutils` to write the pci names. This can
have unexpected consequences, and in our case was causing
(rare) segfaults.

As recommended also in [ctypes FAQ](https://github.com/ocamllabs/ocaml-ctypes/wiki/FAQ#-whats-the-best-way-to-bind-c-functions-that-act-on-strings)
we can solve this issue by allocating a C array (that lives
outside the ocaml heap) and use it as the string buffer
expected by the C function. This is still managed by ocaml
and will be freed by the GC once it is no longer in use
but it will not change unexpectedly under our feet.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>