CA-282012: enable/disable clustering daemon (#3430)
* CA-282012: enable/disable clustering daemon

Starting the cluster daemon on the first XAPI RPC call doesn't work when HA is
enabled and we need to start clustering from the static-vdis script before XAPI
is fully operational.

Instead of starting/stopping the clustering daemon from XAPI on boot let systemd
do it: just enable/disable the service when a cluster host is created/destroyed.

This still keeps the semantics we wanted, i.e. when clustering is disabled we do
not have a running clustering daemon.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>