CA-282740: handle errors from execve in forkexecd
Exit with status 127 when command is not found and 126 when not executable,
as POSIX shells do.

Previously forkexecd claimed that the process exited with code 1 when
the command was not found, which can be very confusing when exit code 1
has a meaning for the command you are running.

E.g. nbd-client check exits with code 1 when the device is not in use,
but if you got the path to nbd-client wrong forkexecd claimed it exited with 1.

Follow POSIX conventions, see
 http://pubs.opengroup.org/onlinepubs/9699919799//utilities/V3_chap02.html#tag_18_08_02

Signed-off-by: Edwin Török <edvin.torok@citrix.com>