CP-26862: Set RLIMIT_FSIZE for QEMU
During a migration, QEMU is given a writable file descriptor to write
its save image. This could allow a denial of service since a rogue QEMU
could fill up the filesystem backing the file descriptor. To prevent
this, set the hard and soft limits of RLIMIT_FSIZE to 32 KiB. This is
enough for QEMU to save its migration state. QEMU is still free to write
to its emulated disks since this limit does not apply to block devices.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>