Distinguish between domain types using xenstore key
Now that we have a third domain type (PV-in-PVH), we need a more reliable way
to determine the domain type of a running VM. The traditional way is to ask Xen
for the domain info and consult the `hvm_guest` key. However, PVH guests are
also HVM guests (to Xen), so we need a better way to distinguish them.

It is possible to distinguish between PVH and HVM by looking at the emulation
flags (also part of the domain info). However, we can't use this to tell
PV-in-PVH and plain PVH guests apart (once we introduce those).

Instead, xenopsd now writes the domain type to
`/local/domain/<domid>/domain-type` in xenstore, on domain create. Whenever and
wherever we need to know the type of the domain, this xenstore key is a
reliable place to find it (for xenopsd, squeezed).

Possible values for the key, at the moment, are "hvm", "pv", and "pv-in-pvh".

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>