CA-283697: Avoid resetting a field to null when creating an object from a Hashtable not containing the field.
Created a new method to update an object from a Hashtable setting only the
fields contained in the Hashtable. Changed the class constructor from Hashtable
to initialise the fields with their default values and only set the ones
contained in the Hashtable instead of resetting them to null.
Also, simplified the methods in the Marshalling class, mostly to make the change
in the generation code easier.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>