[CP-26826] Add fsync on pool join when clustering enabled
We can't add code to `pool-join` that is aware of the state of clustering on the pool master, nor that can perform an fsync on the master - because the pool master isn't correct on the new slave until its host agent restarts. Instead we add an fsync call in the constructor of `Cluster_host` objects, by requesting a database sync which, since clustering is enabled, will also fsync the database.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>