[CP-26826] Fix preferred_write_db to remove List.Hd
`preferred_write_db` uses List.Hd to get the preferred DB, but there may be no connections (this is the case in some situations in the tests). Therefore these tests would fail if they called the function. It's modified to return an Option, letting callers deal with the None case. Also remove the unneeded unit parameter.

Where a caller is already catching + handling errors, they'd already be dealing with this List.Hd error if it occured so we use `Opt.unbox` which raises Not_found in the None case. In `xapi_pool_transition.ml` we use the DB connection as input to `flush_and_exit` ie we expect to exit xapi, so unbox is used to ensure we don't fail silently.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>