CA-283728: put generated code into new namespace
The already shipped supp-pack with the tmpfs SR imports a python library from
the installed host.
If we update the host without updating the supp-pack we need to keep the python
library backward compatible.

The short-term fix is to keep the old library as was, and put the new library
into a different module.
In the future the code generator can be updated to generate a single python
module that is  compatible with all the APIs.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>