CA-283728: backward compatibility for uuid
SR.attach and SR.create take a uuid parameter now.
If we are talking to an old plugin version then take the uuid parameter out of
the JSON that we send it.

Fix SR.scan compatibility:
SR.ls returns an array, we need to fix each element inside that for backward
compatibility.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>