CP-25681: clean up /var/run/xen/qmp-{libxl,event}-* after qemu terminates
As for the vnc socket file, the qmp files qmp-{libxl,event}-<domid> are
created by QEMU. We believe it's better to delete them in xenopsd because
QEMU could crash and leave this file over.

The cleanup is a best-effort approach. A failure is not propagated,
but logged for analysis.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>