CP-25898 add tap_open() primitive
We need to open /dev/net/tun with specific ioctl calls. This could be
implemented by either using a library like

  https://github.com/mirage/ocaml-tuntap

or we can implement the specific use case directly and avoid the new
dependency. This commit is a direct implementation of the required
functionality. It is accessible from OCaml as:

external _tap_open : string -> Unix.file_descr = "stub_tap_open"

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>