CP-26340: SR.probe changes to match xcp-idl
We have 3 categories of probe results depending where you use it:
* attachable: suitable for SR.attach
* creatable: suitable for SR.create
* incomplete: suitable for further SR.probe

This is done by converting the response from SMAPIv3 plugin, which returns an
SR, and a complete boolean flag.
Reject with an error incompatible results
(such as a configuration that is both complete and has an SR,
 a complete configuration without a uuid,
 or an incomplete configuration with a uuid).
"complete" in SMAPIv3 means it is usable for SR.create.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>