Add fe_argv module for incrementally building argv
Module fe_argv helps to build a command line incrementally in a monadic
way. This is most useful when a process to be started with forkexed
requires a long and complicated list of arguments and/or is passed file
descriptors.

This is a purely auxiliary module - it does not interact with the
existing code directly.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>