CP-27201 Sync network sriov from master to slave when slave restarting
For every sriov object in master, find the corresponding PIF in slave as
 following figure shows:
 sriov
   |
 PIF(master)     <-- same device name -->  PIF(slave)

If cannot find, then continue to process next sriov.
If there is an existing sriov on the slave PIF, keep it as is.
Otherwise try to create sriov on the slave PIF and attache the sriov logical
PIF to the same network as sriov on master.

Signed-off-by: Yang Qian <yang.qian@citrix.com>