import, importexport: further reduce the use of XMLRPC and cleanup code
This removes the necessity to pass trough an intermediate re-stringification of
the various datastructures. There is still an open question about the need of
Bigbuffer. If we can cope with strings as done now, we could actually get rif
of Bigbuffer in all the rpc interaction, and maybe in xapi itself. If it is needed
it should not be hard to provide the necessary infrastructure in ocaml-rpc to
deal with it. We could also probably rely on Xmlrpc.from_a although I am a bit
wary of that call...

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>