CP-27329: Make Network SR-IOV Sync Do Plug on Slave
During reboot of a slave XenServer in a pool, XAPI logic on slave will
try to synchronize some networks from master. These networks include
bond, sriov, vlan, and tunnel. There is a difference between bond/sriov
and vlan/tunnel: the sync of bond/sriov is done on slave host, while the
sync of vlan/tunnel is done on master host. This is because sync of
vlan/tunnel is just to create DB objects without plug. Making it done
on master could improve the efficiency of creating DB objects.

But this is not true for bond and sriov, as both of them need to be
plugged in sync, which has to be done on slave.

This change will fix a bug which would cause the plug of sriov done on
master. Instead, it should be done on slave, just like sync of bond
network.

Signed-off-by: Ming Lu <ming.lu@citrix.com>