Add backwards-compatible SR.probe output for SMAPIv3
For SMAPIv1, SR.probe just returns an XML blob. This commit reconstructs
an XML with the same shape for SMAPIv3, where we get a data structure
instead, to achieve backwards-compatibility for API users that expect
this XML structure to be returned from SR.probe. Previously we just
converted this data structure into XML, which resulted in a completely
different XML document.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>
Signed-off-by: Edwin Török <edvin.torok@citrix.com>