CA-285751: Fix VIF.get_state to return Vif.active as reality
The issue to be fixed in this commit is that SR-IOV VF backed vif
didn't return the real Vif.active when the SR-IOV VF (PCI) is not
plugged. This makes XAPI think this VIF should be removed and set its
currently_attached as false in a rebooting scenario. Consequently, XAPI
would find the metadata of VM changed, and push new memtadata to
Xenopsd, during which XAPI filtered out the un-attached vifs in new
metadata. Finally, this vif would not be created in the guest.

The fix is to make VIF.get_state return real value of active, nomorally
it is true, so that XAPI will not set this vif as non-attached.

Signed-off-by: Ming Lu <ming.lu@citrix.com>