CA-285897: Filter out un-attached VF-backed vifs in PCI metadata
The background logic is that XAPI will try to construct all information
on a VM as metadata and push this metadata to Xenopsd to manage the VM.
Particularly the logic on vifs are: for the `start` of VM, XAPI will try
to push all vifs, no matter attached or not, to Xenopsd; while for the
`reboot` of VM, it will try to push the vifs which are currently
attached only.

SR-IOV VF backed vifs should follow same logic. This means XAPI should
not push the VFs' PCIs which are backend of un-attached vifs also.

This fix adds a filter on contructing PCI metata to filter out the vifs
which are backed by VF PCIs and currently un-attached.

Signed-off-by: Ming Lu <ming.lu@citrix.com>