CA-284768: Improve thread-safety of TypedTable
The `TypedTable` functor is used for several "databases" in xenopsd, and
contains a set of functions to read and write metadata to a filesystem. The
current functions are thread-safe with respect to the low-level file system
operations, which prevent corruption. However, certain uses of the `read` and
`write` functions in concurrent threads may still interfere.

A pattern that is used in several places in the Xenops_server_xen backend is to
read from the DB, modify the value, and write it back. If this sequence isn't
protected by a mutex, then it is possible for a second thread to write to the
same key after the read but before the write of the first thread. The write
from the second thread will then be undone by the first.

To protect against this, an `update` function is introduced, that does an
atomic read-modify-write, which should be used in place of read+write. Also the
existing `add` and `remove` functions are protected by the same mutex. The
`delete` and `write` functions are not, and should be avoided.

Note that `update` can always be used in place of `write` or `delete`. Also
`add` and `remove` can likely be removed. However, they are heavily used in
Xenops_server, which we are not changing immediately.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>