CA-284768: Use DB.update_exn in VM.build
The bug exposed itself where the read-modify-write cycle in `VM.build`
interfered with a similar cycle in `store_rtc_timeoffset` (fixed in the
following patch) that happened at the same time.

In `VM.build`, we use `DB.update_exn`, which is like `DB.update`, but it raises
an exception if the read fails. This is to match the behaviour of the original
code that uses `DB.read_exn`.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>