CA-285511: TCUSBPassBetweenVMWithinPool: No route to host
This error is caused by calling unplug vusb which will call usb_reset to
remove device from cgroup allow list. The root cause is that we didn't
filter vusb when creating metadata according vusb field `currently_attached`.
So if we create vusb by CLI when vm is running and then shutdown the vm, this
error will be raised by the usb_reset.py with "No such file or directory",
as the vusb is not plugged to vm and then the file is not created.

So we need to filter the vusb if currently_attached is false, it
shouldn't be pushed to xenopsd.

Signed-off-by: Taoyong Ding <taoyong.ding@citrix.com>