CA-281178: Update allowed operations in consider_enabling_host
In a pool, restarting the toolstack on a slave will go through two
stages: first shutting down, and next booting up again.  In the
shutting-down stage, Host.enabled will be set as false. Then in the
booting-up stage, XAPI will set it back to true in the 'initialising
storage' step.

The bug to be fixed in this ticket is that the allowed_operations of
the slave host is not updated correspondingly when setting
Host.enabled back to true while booting up.

Signed-off-by: ming <ming.lu@citrix.com>