CA-286115: Remove race between DB and xenstore watch thread
The xenstore watch @introduceDomain will be emitted whenever a new
domain turns up. We have a thread that reacts to this and installs watches
for the new domain if it's one xenopsd is interested in. The problem occurs
here because xenopsd decides whether a domain is interesting or not based on
whether it has metadata about the VM in its VmExtra database, `DB`. This
was only stored on completion of the VM_create call, which is after the
domain is created, and hence there was a race between the two threads. This
commit ensures there is an entry in the VmExtra database before the
domain is created, thus removing the race.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>