CA-286135: Bring down SR-IOV physical PIF when bring down bond
This fix will add logic to bring down SR-IOV physical PIF if it is also
a slave of bond and this bond is being brought down.

This is because following PIF structure is required to supported:

(Bond slave) -------------------+
                                |
                                +- (Bond master)
                                |
(Bond slave + SR-IOV physical)--+
            |
            |
    (SR-IOV logical)

Meanwhile, this commit also refines the code structure in `unplug` to
put exclusive PIF types into a pattern match.

Signed-off-by: ming <ming.lu@citrix.com>