CA-286338: Inter-host VM copy failed from source host to a slave in a pool
In `Helpers.call_api_functions_internal`, if it is called in slave and the
session is not forwarded from master (e.g. cross pool VM import), then we may
get a session without 'pool' flag set, this behavior is inconsistent with the
comments and the comments is correct.
So when we get the session from context, we should always make a new one if
it's not a inner pool session.

Signed-off-by: Yang Qian <yang.qian@citrix.com>