CA-286455: Make lower MMIO hole twice as big for the VMs with vGPU
We extend the default MMIO hole start for the VMs that either start
(there is a companion change in xenguest to support this) or being
migrated in from qemu-trad (those have MMIO hole extended by hvmloader)
or QEMU. This becomes the part of ABI from now on and we need to
destinguish VMs without vGPU (with old default) and vGPU VMs.

We rely on the fact that all the existing vGPU now expose only 256MB
BAR and therefore extend qemu-trad MMIO hole to a known size.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>