HFX-2157 CA-285840: Check for host liveness in the new stunnel hook
In the stunnel connection retry loop a new hook has been added to allow
for a test to be run before going around the loop. This commit populates
the hook with a function to check for host_metrics.live so that if a
host has been marked as dead we exit the retry loop quickly.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>