CP-27560 during xapi upgrade, upgrade VM device model
This commit makes sure, a VM's device model is updated to the new
default qemu-upstream-compat. The commit relies on
Xapi_vm_helpers.ensure_device_model_profile_present setting the correct
default when no device model is present and to upgrade an existing
device model when necessary. This upgrade rule is always applied.

To make sure the xapi upgrade rule is triggered, the minor version of
the xapi database model is bumped.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>