CA-289907: Base pool.cpu_info:features_hvm on HVM-capable hosts only
This field should be the intersection of the HVM-specific feature sets over all
hosts in the pool that can actually run HVM guests.

The `host.cpu_info:features_hvm` field of non-HVM-capable hosts is a string
with just zeros: no features. Therefore, the intersection currently ends up
disabling all features for HVM guests in the entire pool (including for HVM
capable hosts), which is obviously not what we want.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>