HFX-2174 CA-287863 xe vm-shutdown complete the task too early
This is a backport of 7c31dd1de4ca11ba66267a9bcbbb9e3c8a52510c

When master forward a task to slave, slave will mark the task as
completed once finish processing it, however when master finishing the
forwarding, it will also mark the task as completed. Furthermore, when
slave finished the task, that does not mean the task is done, there may
be some stuff that master has to handle, so we have to ensure the task
only complete once the master finish its work.

For the task forwarded from client, which called with
`exec_with_forwarded_task`, we also have to complete it once it finished.

Signed-off-by: Yang Qian <yang.qian@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@citrix.com>