CA-244573: XenMotion fails after previously attempted SXM is interrupted by XAPI restart and vm goes into suspended state
When executing SXM, the time window between dp creation and persistent is large,
it is possible lead to dp leak if xapi restarted in the time window.
Aiming to mitigate the impact of this issue, this change persistents the dp
before we sending it to remote to shrink the time window.

Signed-off-by: Yang Qian <yang.qian@citrix.com>