CA-288281: Cache tapdisk3 pids
Previously we were listing all files in /dev/shm that started with
`vbd3-`, then finding a domid and devid from the filename, then
looking up the pid of the tapdisk from xenstore. Now we use inotify
to monitor /dev/shm and notice the creation and deletion of these
files. We then look up the pid from xenstore when we need it and
remember it. Deletion of the file triggers the removal of the pid
from the pid cache hashtbl.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>