CA-288281: Use mmap/cstruct rather than read to get stats from blktap3
Once we've got the pid of the tapdisk we use that to locate a file in
/dev/shm containing the statistics. Before this commit we were opening
that file, reading the contents (4k) and closing it once per VBD every
five seconds. This commit replaces that with a mmap of the file that
is done when the tapdisk is first noticed. We then read the values out
directly using cstruct rather than with the c bindings. This should
be much more efficient (especially since the files in question are in
/dev/shm, which was done expressly for this purpose!)

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>