CA-287921: nbd_client_manager: track nbd device -> nbd server mapping
Keep track of the mapping of the connected /dev/nbdX NBD devices to the
Unix domain socket of the server and the exportname used for the
connection. This allows sparse_dd to look up the NBD connection details
based on the NBD device it receives, then it can extract the list of
allocated extents to copy from the NBD server using the BLOCK_STATUS NBD
extension.