CA-287928: Use logger for upstream QEMU logging
QEMU's stdout and stderr are connected to a forkexecd child proccess
which is responsible for logging to syslog. This forkexecd child does
not remain running when using qemu-trad since it logs to syslog via its
privileged half. Since the forkexecd child is fairly heavyweight,
replace it with "logger" which does the same thing but uses
substantially less memory.

Unfortunately, to have the logs prefixed with QEMU's pid implies
starting QEMU before logger which requires a fair amount of code
refactoring in the wrapper so that QEMU is started as a subprocess and
logger is exec()ed rather than exec()ing QEMU directly.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>