CA-288635: redo_log: only log WriteField if the value actually changed
The caller could avoid this by reading the field's value before setting
it, however if the caller is a slave that would incur additional latency
due to the RPC roundtrip.
It is more efficient to detect this on the master, and is a more generic
solution than optimizing individual duplicate writes.

In a log from a production system we see repeated writes to the `domid`
field, and repeated attempts to set `power_state` to `Running`.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>