XSI-1120 CA-288411: make the cache work after a toolstack restart
Backport of 55fd33ece4cf4db2cc6a5dfdbb074811aa03fb52

This fixes a regression introduced in the fix for XOP-830,
where `add_caches` got dropped from the startup resynchronization logic.

This resulted in periodic writes to the database, which are especially
bad with redo log (or implicitly with HA) turned on.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@citrix.com>