CA-288635: block_device_io: reduce number of O_DSYNC writes to half in action_write_delta
An `strace` on `block_device_io` revealed that it does two writes for
every DB delta:
```
16:19:46.694320 read(5, "1", 1)         = 1 <0.000071>
16:19:46.694497 lseek(5, 3979507, SEEK_SET) = 3979507 <0.000058>
16:19:46.694662 select(6, [], [5], [], {1, 997905}) = 1 (out [5], left {1, 997903}) <0.000060>
16:19:46.694863 write(5, "0000000000000087DeleteRow00000016VM_guest_metrics00000046OpaqueRef:57771b01-a339-44fc-af99-c0924c1c07a10000000002684997ebdbeacb-6d6b-f64b-5be8-63d14019fc9d", 155) = 155 <0.001192>
16:19:46.696225 select(6, [], [5], [], {1, 996352}) = 1 (out [5], left {1, 996348}) <0.000074>
16:19:46.696454 write(5, "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0", 16) = 16 <0.000882>
16:19:46.697498 write(4, "writedelta|ack_", 15) = 15 <0.000087>
```

The redo log is opened with O_DSYNC, so each write needs to wait for
a round-trip acknowledgement from the storage, which is a very expensive
operation: a typical HDD can't easily do more than ~100 writes/s
(SSDs of course can).
Doing the concatenation in memory should be more efficient than doing it
on disk.

Since these are consecutive writes this optimization is safe. There are
other places in `block_device_io` where multiple writes are done, and
they must be kept separate, e.g. flushing the main DB writes out some
chunks, and then goes back and marks it as valid. This ordering is
important for crash consistency.

`writev` could also be used, but it is not available in any of the
libraries linked directly into XAPI currently.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>